==> analysis/02-experiment-analysis/04-population-level.Rmd <==
---
output: html_document
---

```{r experiment-04-libraries}
library(tidyverse)
library(here)
library(fs)
library(patchwork)
```

```{r experiment-04-read-data}
lineages_file <- here("data", "processed", "lineages.csv")
lineages_raw_df <- read_csv(
  file = lineages_file,
  show_col_types = FALSE
) %>%
  glimpse()
```

```{r experiment-04-create-factors}
lineages_processed_1_df <- lineages_raw_df %>%
  mutate(
    filamented_id = factor(
      x = filamented_id,
      levels = c(FALSE, TRUE),
      labels = c("Not filamented", "Filamented")
    ),
    filamented_at_time = factor(
      x = filamented_at_time,
      levels = c(FALSE, TRUE),
      labels = c("Not filamented", "Filamented")
    ),
    survived = factor(
      x = survived,
      levels = c(FALSE, TRUE),
      labels = c("Not survived", "Survived")
    ),
    cell_status = interaction(
      filamented_id,
      survived,
      sep = " - "
    ) %>%
      as.character() %>%
      as.factor()
  ) %>%
  glimpse()
```

```{r experiment-04-set-plotting-style}
theme_set(
  theme_bw() +
    theme(
      legend.position = "top",
      strip.background = element_blank()
    )
)
```

```{r experiment-04-create-status-over-time-dataset}
status_time_df <- lineages_processed_1_df %>%
  select(experiment_id, id, time) %>%
  group_by(experiment_id) %>%
  group_modify(~ complete(expand(.x, id, time))) %>%
  ungroup() %>%
  left_join(lineages_processed_1_df) %>%
  rename(cell_status_at_time = filamented_at_time) %>%
  fill(time_first, time_last, gfp_first, filamentation_threshold, ds_red_threshold, .direction = "up") %>%
  fill(antibiotic_start_time, antibiotic_end_time, .direction = "down") %>%
  filter(time >= time_first) %>%
  mutate(
    cell_status_at_time = as.character(cell_status_at_time),
    cell_status_at_time = replace_na(cell_status_at_time, "Dead"),
    cell_status_at_time = factor(
      x = cell_status_at_time,
      levels = c("Not filamented", "Filamented", "Dead")
    ),
    time = factor(time)
  ) %>%
  glimpse() %>%
  identity()
```

```{r experiment-04-status-with-dead}
p_status_with_dead <- status_time_df %>%
  ggplot(aes(x = time, fill = cell_status_at_time)) +
  geom_bar(position = "fill", stat = "count", width = 1) +
  geom_vline(aes(xintercept = factor(antibiotic_start_time)), linetype = "dashed") +
  geom_vline(aes(xintercept = factor(antibiotic_end_time)), linetype = "dashed") +
  facet_grid(experiment_id ~ .) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0), labels = scales::percent) +
  theme(
    panel.spacing.y = unit(1, "lines")
  ) +
  geom_text(
    data = data.frame(
      x = c(7.5, 15.5),
      y = c(0.75, 0.25),
      label = c("Start", "End"),
      experiment_id = "Plasmid"
    ),
    mapping = aes(x = x, y = y, label = label),
    size = 6,
    hjust = 0L,
    vjust = 0L,
    colour = "white",
    inherit.aes = FALSE
  ) +
  labs(
    x = "Time (minutes)",
    y = "Percentage of cells",
    fill = "Cell status"
  ) +
  scale_fill_hue(direction = -1, h.start = 90) +
  NULL

p_status_with_dead
```

```{r experiment-04-metrics-over-time}
p_metrics_over_time <- lineages_processed_1_df %>%
  select(experiment_id, cell_status, time, length, gfp, ds_red) %>%
  pivot_longer(
    cols = c(length, gfp, ds_red),
    names_to = "metric"
  ) %>%
  mutate(
    metric = case_when(
      metric == "ds_red" ~ "DsRed",
      metric == "gfp" ~ "GFP",
      metric == "length" ~ "Length"
    ),
    filamentation_threshold = ifelse(metric == "length", filamentation_threshold, NA)
  ) %>%
  group_by(experiment_id, cell_status, time, metric) %>%
  summarise(
    ci = list(mean_cl_normal(value)),
    .groups = "drop"
  ) %>%
  unnest(cols = c(ci)) %>%
  left_join(
    y = lineages_processed_1_df %>%
      select(experiment_id, antibiotic_start_time, antibiotic_end_time, filamentation_threshold) %>%
      distinct(),
    by = c("experiment_id")
  ) %>%
  mutate(
    filamentation_threshold = ifelse(metric == "Length", filamentation_threshold, NA)
  ) %>%
  ggplot(aes(x = time, y = y, ymin = ymin, ymax = ymax, color = cell_status)) +
  geom_vline(aes(xintercept = antibiotic_start_time), linetype = "dashed") +
  geom_vline(aes(xintercept = antibiotic_end_time), linetype = "dashed") +
  geom_hline(aes(yintercept = as.numeric(filamentation_threshold)), linetype = "dashed") +
  geom_smooth(method = "loess") +
  facet_grid(metric ~ experiment_id, scales = "free_y") +
  labs(
    x = "Time (minutes)",
    y = "Value",
    color = "Cell status"
  ) +
  guides(
    color = guide_legend(ncol = 2),
    fill = guide_legend(ncol = 2)
  ) +
  NULL

p_metrics_over_time
```

```{r experiment-04-proportion-living-cells-gfp-by-row}
counts_survived_by_gfp <- status_time_df %>%
  filter(experiment_id == "Plasmid") %>%
  mutate(
    gfp_first = gfp_first / max(gfp_first),
    gfp_first = cut(gfp_first, breaks = seq(0, 1, 0.05))
  ) %>%
  with_groups(
    gfp_first,
    ~ mutate(.x, n_at_gfp = filter(cur_data(), time == antibiotic_start_time) %>% nrow())
  ) %>%
  group_by(time, gfp_first) %>%
  summarise(
    percentage_alive = sum(cell_status_at_time != "Dead") / first(n_at_gfp),
    antibiotic_start_time = first(antibiotic_start_time),
    antibiotic_end_time = first(antibiotic_end_time),
    .groups = "drop"
  ) %>%
  identity()

survived_p1 <- counts_survived_by_gfp %>%
  ggplot(aes(x = time, y = gfp_first, fill = percentage_alive)) +
  geom_tile() +
  geom_vline(aes(xintercept = factor(antibiotic_start_time)), linetype = "dashed") +
  geom_vline(aes(xintercept = factor(antibiotic_end_time)), linetype = "dashed") +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_discrete(expand = c(0, 0)) +
  scale_fill_viridis_c(
    option = "inferno",
    labels = scales::label_percent(),
  ) +
  labs(
    x = "Time (minutes)",
    y = "Proportion of GFP",
    fill = "Percentage of living cells"
  ) +
  theme(
    legend.spacing.x = unit(1, "cm")
  ) +
  guides(
    fill = guide_colorbar(
      barwidth = 8
    )
  ) +
  NULL

survived_p2 <- counts_survived_by_gfp %>%
  filter(time == antibiotic_end_time + 10) %>%
  ggplot(aes(x = percentage_alive, y = gfp_first, fill = percentage_alive)) +
  geom_bar(stat = "identity") +
  scale_fill_viridis_c(
    option = "inferno",
    labels = scales::percent,
    limits = c(min(counts_survived_by_gfp$percentage_alive), max(counts_survived_by_gfp$percentage_alive))
  ) +
  theme_void() +
  theme(
    legend.position = "none"
  ) +
  NULL

p_proportions_by_gfp_row <- (survived_p1 | survived_p2) +
  plot_layout(
    widths = c(8, 1)
  )

p_proportions_by_gfp_row
```

```{r experiment-04-gfp-survival-probability}
step <- 0.04
breaks <- seq(min(lineages_processed_1_df$gfp) - step, max(lineages_processed_1_df$gfp) + step, step)
hist_gfp_control_info <- hist(
  lineages_processed_1_df$gfp,
  breaks = breaks,
  plot = FALSE
)

p_survival_probability_gfp <- lineages_processed_1_df %>%
  filter(experiment_id == "Plasmid", survived == "Survived", time == time_first) %>%
  group_by(filamented_id) %>%
  summarize(
    counts = list(hist(gfp, plot = FALSE, breaks = breaks)$counts)
  ) %>%
  unnest(counts) %>%
  mutate(
    mids = rep(hist_gfp_control_info$mids, 2),
    control_counts = rep(hist_gfp_control_info$counts, times = 2),
    survival_probability = counts / control_counts
  ) %>%
  identity() %>%
  # filter(survival_probability != 0) %>%
  ggplot(aes(x = mids, y = survival_probability, color = filamented_id)) +
  geom_point() +
  geom_smooth(
    se = FALSE,
    size = 0.5,
    linetype = "dashed"
  ) +
  scale_x_continuous() +
  scale_y_continuous(labels = scales::label_percent()) +
  scale_color_hue(direction = -1, h.start = 90) +
  labs(
    x = "Initial GFP",
    y = "Survival probability",
    color = "Cell status"
  )

p_survival_probability_gfp
```

```{r experiment-04-length-survival-probability}
step <- 1
breaks <- seq(min(lineages_processed_1_df$length) - step, max(lineages_processed_1_df$length) + step, step)
hist_length_control_info <- hist(
  lineages_processed_1_df$length,
  breaks = breaks,
  plot = FALSE
)

survival_probability_length <- lineages_processed_1_df %>%
  filter(experiment_id == "Plasmid", survived == "Survived", time == time_first) %>%
  group_by(filamented_id) %>%
  summarize(
    counts = list(hist(length, plot = FALSE, breaks = breaks)$counts)
  ) %>%
  unnest(counts) %>%
  mutate(
    mids = rep(hist_length_control_info$mids, 2),
    control_counts = rep(hist_length_control_info$counts, times = 2),
    survival_probability = counts / control_counts
  ) %>%
  identity()

p_survival_probability_length <- survival_probability_length %>%
  filter(survival_probability != 1) %>%
  ggplot(aes(x = mids, y = survival_probability, color = filamented_id)) +
  geom_point() +
  geom_smooth(
    se = FALSE,
    size = 0.5,
    linetype = "dashed"
  ) +
  scale_x_continuous() +
  scale_y_continuous(labels = scales::label_percent()) +
  scale_color_hue(direction = -1, h.start = 90) +
  labs(
    x = "Initial length",
    y = "Survival probability",
    color = "Cell status"
  ) +
  coord_cartesian(
    xlim = c(FALSE, 120),
    ylim = c(0, 0.1)
  )

p_survival_probability_length
```

```{r}
p_n_divisions <- lineages_processed_1_df |>
  filter(time <= antibiotic_start_time, survived == "Survived") |>
  group_by(experiment_id, id) |>
  summarize(
    n_divisions = sum(division),
    .groups = "drop"
  ) |>
  ggplot(
    aes(
      x = experiment_id,
      y = n_divisions
    )
  ) +
  ggdist::stat_halfeye(
    adjust = 0.5,
    width = 0.6,
    .width = 0,
    justification = -0.3,
    point_colour = NA
  ) +
  geom_boxplot(
    width = 0.15,
    outlier.shape = NA
  ) +
  gghalves::geom_half_point(
    side = "l",
    range_scale = 0.2,
    alpha = 1 / 30
  ) +
  ggpubr::stat_compare_means(
    method = "t.test",
    comparisons = list(c("Plasmid", "Chromosome"))
  ) +
  scale_y_continuous(
    breaks = 0:6
  ) +
  coord_flip() +
  theme_minimal() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor.y = element_blank(),
    panel.grid.minor.x = element_blank(),
    legend.position = "none",
    plot.title.position = "plot",
    plot.title = element_text(
      face = "bold",
      size = 14,
      hjust = 0
    )
  ) +
  labs(
    x = "",
    y = "Number of divisions before exposure to antibiotics",
    title = "Survival cells"
  ) +
  NULL

p_n_divisions
```

```{r experiment-04-dt-length-survival-probability}
complete_time_dt_stats <- lineages_processed_1_df |>
  arrange(experiment_id, id, filamented_id, time) |>
  add_count(experiment_id, id, filamented_id, wt = division) |>
  filter(n != 0) |>
  group_by(experiment_id, id, filamented_id) |>
  summarize(
    mean_dt_length = mean(lag(length), na.rm = TRUE),
    .groups = "drop"
  ) |>
  identity()


survival_time_dt_stats <- lineages_processed_1_df %>%
  add_count(experiment_id, id, filamented_id, wt = division) |>
  filter(
    n != 0,
    survived == "Survived",
    time <= antibiotic_start_time
  ) %>%
  arrange(experiment_id, id, filamented_id, time) |>
  group_by(experiment_id, id, filamented_id) |>
  summarize(
    mean_dt_length = mean(lag(length), na.rm = TRUE),
    .groups = "drop"
  ) |>
  identity()
```

```{r do-it-for-chromosome}
chromosome_complete_dt_length <- complete_time_dt_stats |>
  filter(experiment_id == "Chromosome")

chromosome_survived_dt_length <- survival_time_dt_stats |>
  filter(experiment_id == "Chromosome")

step <- 5
breaks <- seq(
  min(chromosome_complete_dt_length$mean_dt_length) - step,
  max(chromosome_complete_dt_length$mean_dt_length) + step,
  step
)

chromosome_hist_dt_length_control_info <- hist(
  chromosome_complete_dt_length$mean_dt_length,
  breaks = breaks,
  plot = FALSE
)

chromosome_survival_probability_dt <- chromosome_survived_dt_length |>
  group_by(experiment_id) |>
  summarize(
    counts = hist(mean_dt_length, plot = FALSE, breaks = breaks) |>
      pluck("counts") |>
      list()
  ) |>
  unnest(counts) |>
  mutate(
    mids = chromosome_hist_dt_length_control_info$mids,
    control_counts = chromosome_hist_dt_length_control_info$counts,
    survival_probability = counts / control_counts
  ) %>%
  glimpse()



p_survival_probability_chromosome_dt_length <- chromosome_survival_probability_dt %>%
  filter(survival_probability != 1) %>%
  ggplot(aes(x = mids, y = survival_probability)) +
  geom_point() +
  geom_smooth(
    se = FALSE,
    size = 0.5,
    linetype = "dashed"
  ) +
  scale_x_continuous() +
  scale_y_continuous(labels = scales::label_percent()) +
  labs(
    x = "Mean length lag difference",
    y = "Survival probability"
  )

p_survival_probability_chromosome_dt_length
```

```{r do-it-for-plasmid}
plasmid_complete_dt_length <- complete_time_dt_stats |>
  filter(experiment_id == "Plasmid")

plasmid_survived_dt_length <- survival_time_dt_stats |>
  filter(experiment_id == "Plasmid")

step <- 5
breaks <- seq(
  min(plasmid_complete_dt_length$mean_dt_length) - step,
  max(plasmid_complete_dt_length$mean_dt_length) + step,
  step
)

plasmid_hist_dt_length_control_info <- hist(
  plasmid_complete_dt_length$mean_dt_length,
  breaks = breaks,
  plot = FALSE
)

plasmid_survival_probability_dt <- plasmid_survived_dt_length |>
  group_by(experiment_id) |>
  summarize(
    counts = hist(mean_dt_length, plot = FALSE, breaks = breaks) |>
      pluck("counts") |>
      list()
  ) |>
  unnest(counts) |>
  mutate(
    mids = plasmid_hist_dt_length_control_info$mids,
    control_counts = plasmid_hist_dt_length_control_info$counts,
    survival_probability = counts / control_counts
  ) %>%
  glimpse()

p_survival_probability_plasmid_dt_length <- plasmid_survival_probability_dt %>%
  filter(survival_probability != 1) %>%
  ggplot(aes(x = mids, y = survival_probability)) +
  geom_point() +
  geom_smooth(
    se = FALSE,
    size = 0.5,
    linetype = "dashed"
  ) +
  scale_x_continuous() +
  scale_y_continuous(labels = scales::label_percent()) +
  labs(
    x = "Mean length lag difference",
    y = "Survival probability"
  )

p_survival_probability_plasmid_dt_length
```

```{r other-option}
lineages_processed_1_df |>
  group_by(experiment_id, id, cell_status, antibiotic_start_time) |>
  summarize(
    time = time,
    dt_length = lag(length, n = 1),
    .groups = "drop"
  ) |>
  drop_na() |>
  filter(time <= antibiotic_start_time) |>
  ggplot(aes(x = factor(time), y = dt_length, fill = cell_status)) +
  geom_boxplot(outlier.shape = NA) +
  facet_grid(~experiment_id) +
  coord_cartesian(ylim = c(0, 90)) +
  labs(
    x = "Time",
    y = "Lenght dt",
    fill = "Cell status"
  )
```

```{r}
figure_to_save_dt <- lineages_processed_1_df %>% 
  select(experiment_id,
         id,
         survived,
         length,
         time,
         antibiotic_start_time,
         time_last
  ) %>% 
  filter(time <= time_last, time < antibiotic_start_time) %>% 
  arrange(experiment_id
          , id
          , time
  ) %>% 
  group_by(experiment_id,
           id,
           survived
  ) %>% 
  mutate(
    growth_rate = (length - lag(length)) / lag(length) * 100
  ) %>% 
  summarize(
    mean_growth_rate = mean(growth_rate, na.rm = TRUE),
    .groups = "drop"
  ) %>% 
  ggplot(aes(x = survived, y = mean_growth_rate, fill = survived)) +
  geom_boxplot() +
  ggpubr::stat_compare_means(
    method = "t.test",
    comparisons = list(c("Survived", "Not survived"))
  ) +
  facet_grid(. ~ experiment_id) +
  coord_flip() +
  labs(
    x = "Cell status",
    y = "Mean growth rate ( (present - past) / past * 100 )"
  ) +
  theme(legend.position = "none") +
  NULL

figure_to_save_dt
```

```{r}
lineages_processed_1_df %>% 
  select(experiment_id,
         id,
         survived,
         length,
         time,
         antibiotic_start_time,
         time_last
  ) %>% 
  filter(time <= time_last, time < antibiotic_start_time) %>%
  arrange(experiment_id
          , id
          , time
  ) %>% 
  group_by(experiment_id,
           id,
           survived
  ) %>% 
  mutate(
    growth_rate = (length - lag(length)) / lag(length) * 100
  ) %>% 
  summarize(
    mean_growth_rate = mean(growth_rate, na.rm = TRUE),
    .groups = "drop"
  ) %>% 
  drop_na() %>% 
  group_by(
    experiment_id,
    survived
  ) %>% 
  summarise(
    what = 'Mean growth rate',
    min = min(mean_growth_rate),
    median = median(mean_growth_rate),
    mean = mean(mean_growth_rate),
    max = max(mean_growth_rate),
    std = sd(mean_growth_rate)
  )

```

```{r}
figure_to_save <- lineages_processed_1_df %>% 
  filter(time <= antibiotic_start_time) %>% 
  group_by(experiment_id, id) %>% 
  summarize(
    n_divisions = sum(division)
  ) %>% 
  ggplot(aes(x = experiment_id, y = n_divisions)) +
  stat_eye(position = "dodge") +
  stat_compare_means(
    label = "p.format",
    method = "t.test",
    #ref.group = "Survived",
    label.x.npc = 0.75,
    hide.ns = TRUE
  ) +
  theme(
    panel.grid.major = element_blank(),
    legend.position = "none"
  ) +
  labs(
    x = "Cell status",
    y = "Number of divisions prior to drug introduction",
    fill = "Cell status"
  ) +
  coord_flip() +
  NULL

figure_to_save

lineages_processed_1_df %>% 
  filter(time <= antibiotic_start_time) %>% 
  group_by(experiment_id, id) %>% 
  summarize(
    n_divisions = sum(division)
  ) %>% 
  group_by(experiment_id) %>% 
  summarize(
    what = 'Number of divisions prior to drug introduction',
    min = min(n_divisions),
    median = median(n_divisions),
    mean = mean(n_divisions),
    max = max(n_divisions),
    std = sd(n_divisions)
  )
```

```{r}
lineages_processed_1_df %>% 
  filter(time < antibiotic_start_time) %>% 
  group_by(experiment_id, survived, id) %>% 
  summarize(
    n_divisions = sum(division)
  ) %>% 
  group_by(experiment_id, survived) %>% 
  summarize(
    what = 'Number divisions prior to drug introduction',
    min = min(n_divisions),
    median = median(n_divisions),
    mean = mean(n_divisions),
    max = max(n_divisions),
    std = sd(n_divisions)
  )
```
